pyproject.toml
 [build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit_persona"
version = "0.1.0"
description = "A user persona developing module based on the posts and comments from reddit."
authors = [{ name = "Guru Hemendra", email = "guruhemendraputhuru@gmail.com" }]
readme = "README.md"
requires-python = "==3.11.3"
dependencies = [
    "praw>=7.0.0",                      # For interacting with Reddit
    "textblob>=0.15.0",                  # For text processing
    "chromadb>=0.1.0",                   # For vector storage
    "sentence_transformers>=2.0.0",      # For sentence embeddings
    "transformers>=4.0.0",               # Hugging Face Transformers
    "huggingface-hub>=0.10.0",            # Hugging Face Hub integration
    "langchain_huggingface>=0.0.5",      # LangChain integration with Hugging Face
    "PyYaml",
    "torch",
    "langchain-community>=0.2.0",        # Ensure this is the latest version of langchain-community
]
